Don't try replace by reference when @ is in the middle of text
